Schema split and code formatting cleanup
The relatively lengthy schema has been split into a separate file, and
black and flake8 formatting has been enforced.

update gitignore to template

poetry and dockerfile update
